requirements.txt
anthropic>=0.40.0
PyGithub>=2.3.0
PyYAML>=6.0
pytest>=8.0
